(chore) move clickhouse to correct directory
